L1Rate.cc file to skim and set a flag from ZB Ntuple for events passing specific L1 seed

+++ NtupleMaker/python/trigger_tree.h
@@ -14,7 +14,7 @@ public :
     // from HLT
     int runNumber;
     int lumiBlock;
-    int eventNumberID;
+    unsigned long long eventNumberID; // sometimes needs to be int for older samples...
 
     int nEvents;
